backend/requirements.txt
Django==6.0.3
djangorestframework==3.15.2
django-cors-headers==4.7.0
python-dotenv==1.0.1
google-generativeai==0.8.4
psycopg2-binary==2.9.11
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.8.2
